requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
langgraph>=0.1.0
qdrant-client>=1.6.0
openai>=1.0.0
redis>=5.0.0
rq>=1.0.0
pyjwt>=2.8.0